mutation_reader: gallop mode for combined reader
In case when a single reader contributes a stream of fragments
and keeps winning over other readers, mutation_reader_merger will
enter gallop mode, in which it is assumed that the reader will keep
winning over other readers. Currently, a reader needs to contribute
3 fragments to enter that mode.

In gallop mode, fragments returned by the galloping reader will be
compared with the best fragment from _reader_heap or _fragment_heap.
If it wins, the fragment is directly returned. Otherwise, gallop mode
ends and merging performed as in general case, which involves heap
operations.

Fixes #3593.