cdc: produce metadata on generated mutations in augment_mutation_call
Adds a second return value for augment_mutation_call, which is an object
that describes how mutations that were produced due to CDC augmentation
are laid out in the result vector of mutations. This object will be
useful in storage_proxy for tracking groups of CDC log mutations that
resulted from a single base mutation.