storage_proxy: add cdc_augmentation_tracker
Adds a cdc_augmentation_tracker, an object that is responsible for
tracking the result of writing CDC log mutations coming from a single
base mutation. If writing of any of the mutations from the group fails,
cdc_augmentation_tracker will increase metrics related to failed CDC
augmentation operations.

Actual registration of the tracker will be done in subsequent patches.