hinted handoff: don't keep replay_position for too old hints
The send_one_file_ctx::rps_set keeps track commitlog positions of hints
that are currently being sent. If sending of one hint fails, its
position is not removed from rps_set.

If some hints fail to be sent, rps_set is used to calculate commitlog
position of the first failed hint. Sending of hints from the same file
is attempted again after some delay, and is started from the saved
position.

Additionally, if a hint is older than gc_grace_seconds + hint flush
period, it is silently dropped. In this case, due to a bug, the
commitlog position was inserted into rps_set but not removed after
discarding the hint. This would cause rps_set to grow large if a hint
file contained many old hints.

This patch fixes send_one_hint function so that it removes hint's
commitlog position from rps_set when it is being dropped due to old age.