hinted handoff: always create hints manager
Up to this point, hints_manager for regular hints was not created if
hinted_handoff_enabled option was set to "false". Because of that, the
hints_manager for regular hints cannot react to configuration change if
hints were disabled at boot time.

This commit causes hints_manager for regular hints to always be created.

This introduces an important behavioral change - if Scylla had hints
enabled, was stopped when there were some hints on disk and then later
restarted with hints disabled, this will cause stored hints to be sent
out. Before this change, this scenario would not cause those hints to be
sent.