config: add wait_for_hint_replay_before_repair_in_ms option
Adds the `wait_for_hint_replay_before_repair_in_ms` configuration
option. It specifies how long the repair coordinator will wait for
hints from nodes participating in repair to be replayed. It is 0 by
default, which means that the repair coordinator will not wait for
hints at all before continuing with repair. This option is
live-updateable. It will be used in subsequent commits from the same
PR.