Revert "repair: plug in waiting for hints to be sent before repair"
This reverts commit 49f4a2f9689860b728c10284eb7ccc79743a2068.

The idea to wait for hints to be replayed before repair is not always a
good one. For example, someone might want to repair a small token range
or just one table - but hinted handoff cannot selectively replay hints
like this.

The fact that we are waiting for hints before repair caused a small
number of regressions (#8612, #8831).

This commit removes the logic in repair which caused it to wait for
hints.