Revert "config: add wait_for_hint_replay_before_repair option"
This reverts commit 86d831b319c487dbf6da5f3f04935b397b018de0.

This commit removes the wait_for_hints_before_repair option. Because a
previous commit in this series removes the logic from repair which
caused it to wait for hints to be replayed, this option is now useless.

We can safely remove this option because it is not present in any
release yet.