hints: get rid of move constructors in ep manager and sender
Both `end_point_hints_manager` and `sender` had their some of the
move/copy constructors defined, but it doesn't make sense to move them.
The end_point_hints_manager is put into an unordered_map and never moved
again, and sender is just a field of the end_point_hints_manager.

The move constructor of the endpoint manager was used in a single place,
namely when the endpoint manager was inserted into the _ep_managers map
with the unordered_map::emplace method, like this:

    ep_managers.emplace(ep, end_point_hints_maanger(ep, *this))

The move constructor was needed because `emplace` forwards its argument
to the std::tuple<const Key, Value> constructor which takes Value as a
rvalue reference and moves it.

The unordered_map::emplace method was changed to
unordered_map::try_emplace - now, its arguments are forwarded directly
to the end_point_hints_manager's constructor and the move constructor is
not needed.