Revert "db/hints: allow to forcefully update segment list on flush"
This reverts commit e48739a6da82e718fae7927181ba464f5de11231.

This commit removes the functionality from endpoint hints manager which
allowed to flush hints immediately and forcefully update the list of
segments to replay.

The new implementation of waiting for hints will be based on replay
positions returned by the commitlog API and it won't be necessary to
forcefully update the segment list when creating a sync point.