memtable: add id
Adds an ID (UUID) to memtables so that they can be identified inside
Scylla. It will be used by hints code to identify which streaming
memtables were flushed (successfully or not) and to which memtable a
given hint mutation was applied. This tracking will, in turn, be used to
determine which hint files can be deleted on the sender side.