db: add rate_limiter
The rate_limiter is a data structure meant for enforcing per-partition
limits with acceptable level of precision. There is supposed to be one
instance of this structure on each shard.

The algorithm used to track per-partition limits is based on the "lossy
counting" algorithm. Reads/writes are measured over one second windows,
and are recorded in the rate_limiter structure which is a hashmap. At
the end of each time window, the hashmap is cleared and counting starts
again. Additionally, every `W` events, all counters in the hashmap are
decremented by one. This causes the partition counts to be reported
lower than in reality (lower by 1 for every `W` events), but it lowers
the space requirement. Both clearing and decrementing are done in
a lazy fashion.

The parameters were tuned assuming the worst case of 200k operations
per second per shard. I chose `W`=10k, so after every 10k operations
reported read/write counts are lowered by 1 - in the worst case we will
admit 20 requests per-partition over the limit. The resulting size
of the structure is about 1.0MB.

TODO: Due to the fact that measurement is done in windows, the
structure has a tendency of accepting requests to the rate-limited
partitions in bursts, i.e. if the limit is X writes/s but a partition
receives 100X writes/s, then X requests are accepted at the beginning
of each measuring window, and this amount may drown other requests if
they are expensive. We should consider modifying the structure so that
it accepts/rejects writes to a hot partition in a more spread-out
fashion.