cql_test_env: throw exceptions passed as result_message::exception in execute_cql
Usually, a `result_message` is processed by using a visitor. In the
previous commit, existing visitors were adjusted so that if they
encounter a `result_message::exception` they will throw the exception
held in it. However, some of the methods in `single_node_cql_env` just
return the `result_message` without inspecting it, and the call sites
sometimes just ignore the result, which can lead to silently ignored
exceptions.

This commit modifies the `single_node_cql_env::execute_xyz` family of
methods so that they throw the exception contained in the
`result_message`, if there is any.

As this is purely testing logic, the performance gained from handling
the exception as a value is not as important. Moreover, the test code
was written expecting exceptions to be passed by throw or exceptional
futures, so this brings back the old behavior.