replica: add exceptions
Introduces a set of standard replica-side exceptions. They are meant to
classify all exceptions which can arise on the replica side into a fixed
set of exceptions so that the information about the failure can be
transferred through RPC and interpreted in a meaningful way by the
coordinator.

The set currently consists of:

- unknown_exception: a catch-all exception which covers all exceptions
  not handled by other types in the set. Contains the exception string
  but otherwise type-erases the original exception.
- timeout_exception: a timeout occurred on the replica side.
- forward_exception: a write operation failed to forward mutation.
- virtual_table_update_exception: an update operation on a virtual table
  failed. Contains a string describing the reason of the failure.

As the above are classes derived from std::exception, they can be thrown
directly in the replica code. Hower, it's also perfectly fine to use
other exceptions on the replica side and let them be converted just
before sending them to the coordinator.

Encoding happens in replica::encode_replica_exception, where exceptions
are converted to an exception_variant - a structure holding one of the
standard replica exception types. The exception_variant is serializable
and is supposed to be send/returned over RPC.

The classification above is not exhaustive and more exceptions can be
added in the future. More fields can be added to existing exception
types, too. The encoding allows for both types of extensions without
having to use cluster features.