gms: add TYPED_ERRORS_IN_READ_RPC cluster feature
While the existing code for handling write failures is
backwards-compatible - old coordinators will learn about an error but
won't know the exact type - in case of reads we will have to introduce a
feature. This commit adds TYPED_ERROR_IN_READ_RPC cluster feature which,
when enabled, means that the cluster is ready to return typed errors
from replicas to coordinators.