feature_service: introduce from_feature_set
A list of features for a given node is usually serialized as a
comma-separated string, but our code usually decodes it to a std::set of
strings. There is a function `to_feature_set` which converts the former
to the latter, and this commit introduced `from_feature_set` which
performs conversion in the other direction.

The new method has two overloads: one which takes a `std::set<sstring>`,
and the other one which takes a `std::set<std::string_view>`.

The newly introduced method will be useful in later commits.