storage_service: insert/update supported feature set on start
Each node is responsible for keeping its features up-to-date in the
topology table. When the node joins the cluster or replaces another
node, insert the features along with the join/replace request and
remaining information about the node. If the node just restarts, it will
check and update its features if necessary in
`update_topology_with_local_metadata`.