storage_service/gossiper: don't enable features from gossip if using raft topology coordinator
If raft-based cluster feature management is turned on, then we don't
want to enable features in reaction to what happens in gossip (we still
want to advertise them so that nodes that don't support the new
procedures can still see them).

In this implementation, the node will check whether topology changes
over raft are enabled at startup and will call disconnect_from_gossip
if needed. However, if a proper migration procedure to raft topology
coordinator is implemented, this function should be called right after
the node has switched to raft-based cluster features.