feature_service: add a test-only cluster feature
Adds a cluster feature called TEST_ONLY_FEATURE. It can only be marked
as supported via error injection "features_enable_test_feature", which
should be enabled on node startup via CLI option or YAML configuration.

The purpose of this cluster feature is to simulate upgrading a node to a
version that supports a new feature, and using that to write tests for
the cluster feature logic, e.g. feature becomes enabled only after all
nodes support it, a new node can't join if it doesn't support enabled
feature, etc.

The fact that TEST_ONLY_FEATURE can only be enabled via error injection
should make it impossible to accidentally enable it in release mode and,
consequently, in production.