feature_service::enable: add without_persisting tag
Before #13172, the `feature_service::enable` function was _the_ function
used to enable features. Depending on whether the global db::qctx was
set, that function could persist features by itself or not. After those
changes, the responsibility for persisting features was moved to the new
`persistent_feature_enabler`, which is internal to the feature service,
and `feature_service::enable` now never writes anything to disk.

For the purpose of features over raft, it will be useful to be able to
persist+enable features from the outside. However, enabling features
without persisting them first has its uses in tests and tools.

This commit adds a `without_persisting` tag argument to the
`feature_service::enable` function in order to:

- Make its non-persisting behavior more explicit,
- Differentiate it from an upcoming overload which _does_ persist
  features on disk.