storage_service: reflect raft's supported_features in system.peers
Cluster features are reflected in the `system.peers` table. Currently,
they updated based on gossip. With cluster feature management based on
raft, the group 0 state becomes the source of truth and we need to
update the `system.peers` table when applying commands.

While the logic that updates the features based on gossip is currently
kept active even in raft cluster feature mode, it will eventually be
disabled, so this commit is also a preparation for this.