tests: mark test_blocked_bootstrap as skipped
With the new procedure to join nodes, testing the scenario in
`test_blocked_bootstrap` becomes very tricky. To recap, the test does
the following:

- Starts a 3-node cluster,
- Shuts down node 1,
- Tries to replace node 1 with node 4, but an error injection is
  triggered which causes node 4 to fail after it joins group 0. Note
  that pre-JOIN_NODE handshake, this would only result in node 4 being
  added to group 0 config, but no modification to the group 0 state
  itself is being done - the joining node is supposed to write a request
  to join.
- Tries to replace node 1 again with node 5, which should succeed.

The bug that this regression test was supposed to check for was that
node 5 would try to resolve all IPs of nodes added to group 0 config.
Because node 4 shuts down before advertising itself in gossip, the node
5 would get stuck.

The new procedure to join group 0 complicates the situation because a
request to join is written first to group 0 and only then the topology
coordinator modifies the group 0 config. It is possible to add an error
injection to the topology coordinator code so that it doesn't change the
group 0 state and proceeds with bootstrapping the node, but it will only
get stuck trying to add the node. If node 5 tries to join in the
meantime, the topology coordinator may switch to it and try to bootstrap
it instead, but this is basically a 50% chance because it depends on the
order of node 4 and node 5's host IDs in the topology_state_machine
struct.

It should be possible to fix the test with error recovery, but until
then it is marked as skipped.